Pass All MaterialRepo Tests
The bug is still there: Front-end cannot deal with Tags for some
reason. The result count becomes 0 as soon as MaterialRepo trys to
filter for Tags.

The fact that MaterialRepo is able to Search for an empty collction of tags implies that the bug arieses somewhere else. (Note implies).

SearchProvider, SearchProxy, and MaterialController are also prime suspsects for this bug.