Fix/non string search filters (#86)
* Solved one problem met another

* Docker in progress

* Made a run script

* Change controller route to api/*

* Implemented getting tags from api in searchComponent

* Implemented getting data from API

* Remove tricksy mergeconflict marker

* Add Offset and Limit to Client URI Request

Currently hard-coded for testing. Should be done dynamically for paging.

* Known Good Commit: Search For SearchString

Nukes the rest of the search functionality in MaterialRepo.

* Add WriteLines

* Remove WriteLines + Add MaterialRepo Test

* Pass All MaterialRepo Tests

The bug is still there: Front-end cannot deal with Tags for some
reason. The result count becomes 0 as soon as MaterialRepo trys to
filter for Tags.

The fact that MaterialRepo is able to Search for an empty collction of tags implies that the bug arieses somewhere else. (Note implies).

SearchProvider, SearchProxy, and MaterialController are also prime suspsects for this bug.

* Fix Searching for Tags that Do not Exist + Bump Client Offset to 100

* Add Test For Nonexistent Tags + Fix MaterialController Tests

* Remove Console.WriteLines

* Remove Server.Tests/ConnectionStringTests.cs

* Frankenstein the ConnectionStringTests.cs File

* Remove run.sh

Co-authored-by: Karl Bonde Torp <kart@itu.dk>
Co-authored-by: Emil Østergaard <emil.oestergaard@theorg.com>